removed all usages of automapper, good riddance added a cloningService interface added providers for common factories, can utilize those later to reduce constructor clutter all serializable types can be created directly from the type they serialize all entity types now have a constructor that takes it's serialized type with some additional services to help reverse serialization all entity types have a serialTransform service that contains the required services needed to convert an entity from it's serialized form, and will perform that job serializable types now have public parameterless constructors and init properties to facilitate deserialization many methods from MapInstance have been moved to the respective entities they acted upon the polymorphism in the worldEntity hierarchy is now utilized to determine behaviors added some json converters to the objects in the Chaos.Geometry untangled dependencies caused by the mapper garbage
hmm, not sure if i like Aisling having so many conversions to other objects, I wonder if it would be better to put all of that somewhere else

+++ Chaos/Data/StatSheet.cs
@@ -59,17 +59,17 @@ public record StatSheet : Attributes
     public byte EffectiveDmg => (byte)Math.Clamp(Dmg + DmgMod, byte.MinValue, byte.MaxValue);
     public byte EffectiveHit => (byte)Math.Clamp(Hit + HitMod, byte.MinValue, byte.MaxValue);
     public byte EffectiveInt => (byte)Math.Clamp(Int + IntMod, byte.MinValue, byte.MaxValue);
-    public sbyte EffectiveMagicResistance => (sbyte)Math.Clamp(MagicResistance + MagicResistantMod, sbyte.MinValue, sbyte.MaxValue);
-    public int EffectiveMaximumHp => Math.Max(MaximumHp + MaximumHpMod, 0);
-    public int EffectiveMaximumMp => Math.Max(MaximumMp + MaximumMpMod, 0);
+    public byte EffectiveMagicResistance => (byte)Math.Clamp(MagicResistance + MagicResistantMod, sbyte.MinValue, sbyte.MaxValue);
+    public uint EffectiveMaximumHp => (uint)Math.Max(MaximumHp + MaximumHpMod, 0);
+    public uint EffectiveMaximumMp => (uint)Math.Max(MaximumMp + MaximumMpMod, 0);
     public byte EffectiveStr => (byte)Math.Clamp(Str + StrMod, byte.MinValue, byte.MaxValue);
     public byte EffectiveWis => (byte)Math.Clamp(Wis + WisMod, byte.MinValue, byte.MaxValue);
-    public int HealthPercent => CurrentHp / EffectiveMaximumHp * 100;
+    public int HealthPercent => (int)(CurrentHp / EffectiveMaximumHp * 100);
     public int HitMod => _hitMod;
     public int IntMod => _intMod;
 
     public int MagicResistantMod => _magicResistanceMod;
-    public int ManaPercent => CurrentMp / EffectiveMaximumMp * 100;
+    public int ManaPercent => (int)(CurrentMp / EffectiveMaximumMp * 100);
 
     public static StatSheet Maxed => new()
     {
